requirements.txt
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
pyngrok==6.0.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
urllib3==2.0.3
Werkzeug==2.3.6
